+++ Kokoro/API/Actions.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace Shockah.Kokoro;
 
@@ -11,6 +12,13 @@ public partial interface IKokoroApi
 		CardAction MakeExhaustEntireHandImmediate();
 		CardAction MakePlaySpecificCardFromAnywhere(int cardId, bool showTheCardIfNotInHand = true);
 		CardAction MakePlayRandomCardsFromAnywhere(IEnumerable<int> cardIds, int amount = 1, bool showTheCardIfNotInHand = true);
+
+		CardAction MakeContinue(out Guid id);
+		CardAction MakeContinued(Guid id, CardAction action);
+		IEnumerable<CardAction> MakeContinued(Guid id, IEnumerable<CardAction> action);
+		CardAction MakeStop(out Guid id);
+		CardAction MakeStopped(Guid id, CardAction action);
+		IEnumerable<CardAction> MakeStopped(Guid id, IEnumerable<CardAction> action);
 
 		CardAction MakeHidden(CardAction action, bool showTooltips = false);
 		AVariableHint SetTargetPlayer(AVariableHint action, bool targetPlayer);