+++ Kokoro/I18n.cs
@@ -49,4 +49,9 @@ internal static class I18n
 	public static string StatusEnemyCostActionDescription => "Remove {0} <c=status>{1}</c> from the enemy. If they don't have enough, this action does not happen.";
 
 	public static string EnemyVariableHint => "<c=action>X</c> = The enemy's {0}{1}{2}{3}.";
+
+	public static string ContinueActionName => "CONTINUE";
+	public static string ContinueActionDescription => "Trigger the next actions. If this is not triggered, the next actions will not be either.";
+	public static string StopActionName => "STOP";
+	public static string StopActionDescription => "Stop triggering the next actions. If this is not triggered, the next actions will trigger as usual.";
 }