+++ Kokoro/docs/release-notes.md
@@ -7,6 +7,7 @@
 * For players and modders: Fixed `APlaySpecificCardFromAnywhere` (and in turn `APlayRandomCardsFromAnywhere`) getting into an endless loop if the hand ends up being full in between cards being played (apparent in the Soggins' Do Something! card).
 * For modders: Added `IStatusLogicHook.OnStatusTurnTrigger` and `IStatusLogicHook.HandleStatusTurnAutoStep` APIs.
 * For modders: Added resource cost action APIs.
+* For modders: Added continue/stop action APIs
 * For modders: Added an API to make outgoing X assignment actions.
 
 ## 1.1.1